none

+++ Objetivo1/src/Aluno.java
@@ -30,11 +30,11 @@ public class Aluno {
         this.id = id;
     }
 
-    public int getCpf() {
+    public long getCpf() {
         return cpf;
     }
 
-    public void setCpf(int cpf) {
+    public void setCpf(long cpf) {
         this.cpf = cpf;
     }
